+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [#234] add multifactor management for admin users
+- [#234] enable external ldap configuration
+  - external ldap allows read access
 
 ## [v1.21.0-5] - 2026-06-26
 ### Changed